Add Room

+++ hotel_MVVM/ViewModels/AddRoomViewModel.cs
@@ -1,5 +1,8 @@
 ﻿using hotel_MVVM.Infrastructure.Commands;
 using hotel_MVVM.ViewModels.Base;
+using hotel_MVVM.Views;
+using Interfaces.DTO;
+using Interfaces.Services;
 using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
@@ -8,6 +11,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Interop;
 using System.Windows.Media.Imaging;
@@ -16,6 +20,9 @@ namespace hotel_MVVM.ViewModels
 {
     public class AddRoomViewModel : ViewModel
     {
+        private IRoomService _roomService;
+        private AddAdminWindow _wnd;
+
         private ICommand _addImageCommand;
         private ICommand _addRoomCommand;
         private string _nameImage;
@@ -25,6 +32,12 @@ namespace hotel_MVVM.ViewModels
         private string _roomPrice;
         private string _roomDescription;
         private int _roomPlace;
+
+        public AddRoomViewModel(AddAdminWindow window, IRoomService roomService)
+        {
+            _wnd = window;
+            _roomService = roomService;
+        }
 
         public string RoomName
         {
@@ -92,6 +105,7 @@ namespace hotel_MVVM.ViewModels
             {
                 // Отображаем выбранное изображение в Image элементе
                 _imagePath = openFileDialog.FileName;
+                _nameImage = Path.GetFileName(_imagePath);
                 BitmapImage bitmapImage = new BitmapImage(new System.Uri(_imagePath));
                 SelectedImageSource = bitmapImage;
 
@@ -109,22 +123,38 @@ namespace hotel_MVVM.ViewModels
             {
                 Directory.CreateDirectory(imagesFolder);
             }
-
-            File.Copy(_imagePath, destinationPath, true);
+          
+            if (!File.Exists(destinationPath))
+            {
+                File.Copy(_imagePath, destinationPath, true);
+            }
         }
 
         private void SaveData(object parametr)
         {
+            //int roomPlace = Convert.ToInt32(RoomPlace.SelectedItem);
+            if (!int.TryParse(RoomPrice, out int roomPrice))
+            {
+                MessageBox.Show("Заполните стоимость", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
+                return;
+            }
             if (SelectedImageSource != null)
                 SaveImage();
-            if (string.IsNullOrWhiteSpace(RoomName) || string.IsNullOrWhiteSpace(RoomPrice) ||
+            if (string.IsNullOrWhiteSpace(RoomName) || roomPrice<=0 ||
                  string.IsNullOrWhiteSpace(RoomDesc) || RoomPlace <= 0)
             {
                 MessageBox.Show("Заполните все поля", "Предупреждение", MessageBoxButton.OK, MessageBoxImage.Warning);
                 return; // Прерываем сохранение данных
             }
+            RoomDTO roomDTO = new RoomDTO();
+            roomDTO.RoomName = RoomName;
+            roomDTO.Photo_Path = _nameImage;
+            roomDTO.NumberPlaces = RoomPlace;
+            roomDTO.Price = roomPrice;
+            roomDTO.Description = RoomDesc;
+            _roomService.CreateRoom(roomDTO);
 
-
+            _wnd.Close();
         }
 
         private BitmapImage _selectedImageSource;
